+++ src/app/_features/monitors/pages/create/monitor-create.page.spec.ts
@@ -20,7 +20,7 @@ import { Observable, of } from 'rxjs';
 import { Resource } from 'src/app/_models/resources';
 import { AdditionalSettingsComponent } from '../../components/additional-settings/additional-settings.component';
 import { DurationSecondsPipe } from 'src/app/_shared/pipes/duration-seconds.pipe';
-import { envConfig, EnvironmentConfig } from 'src/app/_services/featureConfig/environmentConfig.service';
+import { envConfig, EnvironmentConfig } from 'src/app/_services/config/environmentConfig.service';
 
 const keyPair = {
   keysandvalues: [
