+++ src/app/_features/visualize/pages/visualize.page/visualize.page.component.spec.ts
@@ -7,7 +7,7 @@ import { ActivatedRoute } from '@angular/router';
 import { MetricsService } from 'src/app/_services/metrics/metrics.service';
 import { SharedModule } from '../../../../_shared/shared.module';
 import { By } from '@angular/platform-browser';
-import { envConfig, EnvironmentConfig } from 'src/app/_services/featureConfig/environmentConfig.service';
+import { envConfig, EnvironmentConfig } from 'src/app/_services/config/environmentConfig.service';
 
 const routes = [
   {
