+++ src/app/_services/monitors/monitor.service.spec.ts
@@ -8,7 +8,7 @@ import { delay } from 'rxjs/operators';
 import { mockResourcesProvider } from 'src/app/_interceptors/request.interceptor';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { APP_INITIALIZER } from '@angular/core';
-import { envConfig, EnvironmentConfig } from '../featureConfig/environmentConfig.service';
+import { envConfig, EnvironmentConfig } from '../config/environmentConfig.service';
 
 describe('MonitorService', () => {
   let injector: TestBed;
