+++ src/app/_services/zones/zone.service.spec.ts
@@ -2,7 +2,7 @@ import { TestBed, getTestBed } from "@angular/core/testing";
 import {  HttpClientModule} from "@angular/common/http";
 import { ZoneService } from "./zones.service";
 import { APP_INITIALIZER } from "@angular/core";
-import { envConfig, EnvironmentConfig } from "../featureConfig/environmentConfig.service";
+import { envConfig, EnvironmentConfig } from "../config/environmentConfig.service";
 
 describe("Zone Service",() =>{
     let injector : TestBed;
