+++ src/app/app.module.ts
@@ -13,7 +13,7 @@ import { SchemaService, AJV_INSTANCE } from './_services/monitors/schema.service
 import { AJV_CLASS, AJV_CONFIG, createAjvInstance } from './_features/monitors/monitors.module';
 import ajv from 'ajv';
 import { AdminAppModule } from 'projects/admin/src/app/app.module';
-import { envConfig, EnvironmentConfig } from './_services/featureConfig/environmentConfig.service';
+import { envConfig, EnvironmentConfig } from './_services/config/environmentConfig.service';
 
 @NgModule({
   schemas: [ CUSTOM_ELEMENTS_SCHEMA ],
